Bab 1 Done

+++ 1. Basic Concepts/simple_operations.py
@@ -1,53 +1,49 @@
 # ================================================== #
 
 # Materi 1 : Simple Operation 1
-print(2 + 2)
-print(5 + 4 - 3)
+print(2 + 2) # 4
+print(5 + 4 - 3) # 6
 
 # Soal 1 : What does this code output ?
 # 1 + 2 + 3
 
 # Jawaban 1 :
-# 6
-print(1 + 2 + 3)
+print(1 + 2 + 3) # 6
 
 # ================================================== #
 
 # Materi 2 : Simple Operation 2
-print(2 * (3 + 4))
-print float(10 / 2)
+print(2 * (3 + 4)) # 14
+print(10 / 2) # 5.0
 
 # Soal 2 : Which option is output by this code ?
 # (4 + 8) / 2
 
 # Jawaban 2 :
-# 6.0
-print float((4 + 8) / 2)
+print((4 + 8) / 2) # 6.0
 
 # ================================================== #
 
 # Materi 3 : Simple Operation 3
-print(-7)
-print((-7 + 2) * (-4))
+print(-7) # -7
+print((-7 + 2) * (-4)) # 20
 
 # Soal 3 : Fill in the blank to make this code correct.
 # (_5 - 1) * 3 = 18
 
 # Jawaban 3 :
-print((-5 - 1) * 3)
+print((-5 - 1) * 3) # -18
 
 # ================================================== #
 
 # Materi 4 : Simple Operation 4
-# print(11 / 0)
-# ZeroDivisionError
+# print(11 / 0) # ZeroDivisionError
 
 # Soal 4 : Drag and drop the answer that will cause a ZeroDivisionError.
 # (17 + 94) / (-5 + _)
 # Choose : 5 or 0 or -5
 
 # Jawaban 4 :
-# print((17 + 94) / (-5 + 5))
-# ZeroDivisionError
+# print((17 + 94) / (-5 + 5)) # ZeroDivisionError
 
 # ================================================== #
